Update all dependencies.
The dependencies changed are the result of running `ncu -ua` [1]. Babel 6 required explicitly specifying the transforms to run on the code. We use the `es2015` preset to get essentially the same functionality as in Babel 5 (or 4). ESLint updates required changes to the configuration [2]. We use the `es6` env to enable all ES6 features except for modules, and the `modules` setting to enable modules.

[1]: https://www.npmjs.com/package/npm-check-updates
[2]: http://eslint.org/docs/user-guide/configuring